+++ server/src/main/java/net/chatfoodie/server/_core/utils/cursor/PageCursor.java
@@ -1,0 +1,7 @@
+package net.chatfoodie.server._core.utils.cursor;
+
+public record PageCursor<T> (
+        CursorRequest nextCursorRequest,
+        T body
+) {
+}
