--- server/src/main/java/net/chatfoodie/server/util/CursorRequest.java
@@ -1,15 +1,0 @@
-package net.chatfoodie.server.util;
-
-public record CursorRequest(
-        Long key,
-        int size
-) {
-    public static final Long NONE_KEY = -1L;
-
-    public Boolean hasKey() {
-        return key != null;
-    }
-    public CursorRequest next(Long key) {
-        return new CursorRequest(key, size);
-    }
-}
