--- server/src/main/java/net/chatfoodie/server/util/PageCursor.java
@@ -1,9 +1,0 @@
-package net.chatfoodie.server.util;
-
-import java.util.List;
-
-public record PageCursor<T> (
-        CursorRequest nextCursorRequest,
-        List<T> body
-) {
-}
