--- server/src/main/java/net/chatfoodie/server/util/PageHelper.java
@@ -1,21 +1,0 @@
-package net.chatfoodie.server.util;
-
-import org.springframework.data.domain.Sort;
-
-import java.util.List;
-
-public class PageHelper {
-    public static String orderBy(Sort sort) {
-        if (sort.isEmpty()) {
-            return "id DESC";
-        }
-        List<Sort.Order> orders = sort.toList();
-
-        var orderBys = orders.stream()
-                .map(order -> order.getProperty() + " " + order.getDirection())
-                .toList();
-
-        return String.join(", ", orderBys);
-    }
-}
-
